+++ src/components/ProductCard/ProductCard.styles.tsx
@@ -1,0 +1,6 @@
+import styled from "@emotion/styled";
+
+export const ButtonWrapper = styled.div`
+  display: flex;
+  justify-content: flex-end;
+`;
