+++ src/components/Spacing/Spacing.styles.tsx
@@ -1,0 +1,5 @@
+import styled from "@emotion/styled";
+
+export const Spacing = styled.div<{ size: number }>`
+  height: ${({ size }) => size}px;
+`;
