+++ src/components/Spacing/Spacing.tsx
@@ -1,0 +1,5 @@
+import * as S from "./Spacing.styles";
+
+export default function Spacing({ size }: { size: number }) {
+  return <S.Spacing size={size} />;
+}
