+++ src/components/index.ts
@@ -3,6 +3,7 @@ export { default as Card } from "./Card/Card";
 export { default as ErrorPopup } from "./ErrorPopup/ErrorPopup";
 export { default as Header } from "./Header/Header";
 export { default as Select } from "./Select/Select";
+export { default as Spacing } from "./Spacing/Spacing";
 export { default as Spinner } from "./Spinner/Spinner";
 export { default as Text } from "./Text/Text";
